add Messaging to demo watch app
