+++ Example/WatchDemo/WatchDemo/WatchDemoApp.swift
@@ -14,6 +14,7 @@
 
 import Firebase
 import SwiftUI
+import FirebaseCore
 
 @main
 struct WatchDemoApp: App {
