+++ Example/WatchDemo/WatchDemoWatchKitApp/Assets.xcassets/AccentColor.colorset/Contents.json
