+++ Example/WatchDemo/WatchDemoWatchKitApp/Assets.xcassets/AppIcon.appiconset/Contents.json
