+++ Example/WatchDemo/WatchDemoWatchKitApp/Assets.xcassets/Contents.json
