+++ Example/WatchDemo/WatchDemoWatchKitExtension/Assets.xcassets/Complication.complicationset/Circular.imageset/Contents.json
