+++ Example/WatchDemo/WatchDemoWatchKitExtension/Assets.xcassets/Complication.complicationset/Graphic Circular.imageset/Contents.json
