+++ Example/WatchDemo/WatchDemoWatchKitExtension/Assets.xcassets/Complication.complicationset/Graphic Corner.imageset/Contents.json
