+++ Example/WatchDemo/WatchDemoWatchKitExtension/Assets.xcassets/Complication.complicationset/Modular.imageset/Contents.json
