+++ Example/WatchDemo/WatchDemoWatchKitExtension/Assets.xcassets/Contents.json
